refactor: use native execs instead of custom execs
